+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   * Dependency Update: eth-tester 0.5.0b1
   * Dependency Update: py-evm 0.3.0a17
   * Protocol: FairSwap with Reusable Smart Contract (#18)
+  * Feature: `environment-info` prints account balances (#21)
   * Fix: Complain about unrecognized parameters
   * Fix: When `-o view True` is given without an output filename use temporary file
   * Fix: Allow slices-count >= 256 in FairSwap contract